Change event queue to use after() instead of sending an event
Sending a custom event from the non-main thread is also not allowed by
Tkinter. This switches the secondary thread -> main thread event queue
to use polling instead as this seems to be the recommended method.
This is another attempt at fixing: "main thread is not in main loop".

Hopefully fixes:
WAHOO-RESULTS-T
WAHOO-RESULTS-W

Signed-off-by: John Strunk <john.strunk@gmail.com>